Remove all unnecessary Gnulib modules
In particular, the various printf-posix modules are not needed as %ls and
the ' flag (thousands separator) is supported on all modern platforms (it
is needed on Cygwin 1.5.x, for example, but the 1.5.x series do not
support swprintf() anyway).  Similarly, OpenBSD 3.8 (released November
2005) is not supported: it does not have <monetary.h>.  A number of
modules (locale, string, unistd, wchar) are only required for NetBSD 5.0
(c. 2009: quite recent).  We try to support MacOS X 10.5 (October 2007),
but it may require GNU libiconv.