Create a README file to describe the new po-translationproject branch
Previously, the po-translationproject branch was a true branch of master,
being merged into master on appropriate occasions.  However, the branch
was never updated to reflect master.

The new branch has been created as an orphan branch, completely empty
apart from this README file.  It will only contain versions of
translations as they appear on the Translation Project website, and will
be merged into master after every update.  No other (public) branch will
ever be merged into this branch.

The last commit on the old branch of the same name was
3553821a5c6a9591bbe3b4795e00f60789d059f0, dated 26th March 2016.