Add strings from obsolete versions of libraries
Previous versions of the GNU Portability Library, and versions of the GNU
C Library as released with various Linux distributions, used strings no
longer present in the current version of the GNU Portability Library.
When used on systems with those strings, Star Traders could not provide
appropriate translations.  This patch fixes that problem.