Remove the Spanish translation by Francisco Javier Serrador
A series of emails with Santiago Vila in May to July 2018 strongly
recommended this option: Francisco's translation is unfortunately of very
poor quality, and although Santiago hoped to provide his own, he has not
been able to do so to date.  A new or significantly revised translation
can always be added at a later date.